Intermittent Speeding Message and Speed Limit Modifier Logic Fix
Testing showed that slack in a multiplayer state showed that intermittent messages of +4 mph from player speed would trigger a player speeding report for 1 tick of DispatcherComms.dll. This was resolved by forcing players to be speeding for 5 seconds before reporting a speeding command. 

The modified Speed Limits (SuperC and Trona) logic was incorrect and adding values together when they should have been set equal to.