Fixes erroneous "0 mph" message spam
Fixes erroneous "0 mph" message bug due to Run8...weirdness. Honestly don't know what's causing this, but it user reports indicate that this happens when coupling into consists. Allegedly it also appears when taking control of a cut of cars, but I haven't been able to find that in testing. Either way, hopefully this will close #11